PDF_processing_module/ESGServer/requirements.txt
easyocr~=1.7.1
pdf2image~=1.17.0
flask~=3.0.2
pillow~=10.2.0